% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contrast.R
\name{Contrast}
\alias{Contrast}
\title{Contrast estimated from an object inheriting from \code{FamModelFit}}
\description{
An R6 class that stores results for a contrast estimated using a
  specified object inheriting from \code{\link{FamModelFit}}. Wald tests and
  confidence intervals for individual rows of the contrast matrix
  \code{L_mat} as well as an overall Wald chi-square test of the null
  hypothesis \code{L_mat \%*\% theta_hat = m} are provided.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Contrast$new()}}
\item \href{#method-get_model_fit}{\code{Contrast$get_model_fit()}}
\item \href{#method-get_L_mat}{\code{Contrast$get_L_mat()}}
\item \href{#method-get_m}{\code{Contrast$get_m()}}
\item \href{#method-get_L_theta_hat}{\code{Contrast$get_L_theta_hat()}}
\item \href{#method-get_V_L_theta_hat}{\code{Contrast$get_V_L_theta_hat()}}
\item \href{#method-get_X2}{\code{Contrast$get_X2()}}
\item \href{#method-get_df_X2}{\code{Contrast$get_df_X2()}}
\item \href{#method-get_p_X2}{\code{Contrast$get_p_X2()}}
\item \href{#method-print}{\code{Contrast$print()}}
\item \href{#method-clone}{\code{Contrast$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Constructs a new instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$new(model_fit, L_mat, m)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_fit}}{An object inheriting from \code{\link{FamModelFit}}.}

\item{\code{L_mat}}{A contrast vector (1 df) or \code{matrix} (>1 df) containing
one contrast in each row.}

\item{\code{m}}{An optional vector containing the null value for each contrast.
Will be set to the zero vector of length \code{nrow(L_mat)} if not
specified.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_model_fit"></a>}}
\subsection{Method \code{get_model_fit()}}{
Returns \code{model_fit} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$get_model_fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_L_mat"></a>}}
\subsection{Method \code{get_L_mat()}}{
Returns \code{L_mat}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$get_L_mat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_m"></a>}}
\subsection{Method \code{get_m()}}{
Returns \code{m}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$get_m()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_L_theta_hat"></a>}}
\subsection{Method \code{get_L_theta_hat()}}{
Returns \code{L \%*\% theta_hat}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$get_L_theta_hat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_V_L_theta_hat"></a>}}
\subsection{Method \code{get_V_L_theta_hat()}}{
Returns \code{L \%*\% V_theta_hat \%*\% t(L)}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$get_V_L_theta_hat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_X2"></a>}}
\subsection{Method \code{get_X2()}}{
Returns Wald chi-square statistic for null hypothesis
  \code{L_mat \%*\% theta_hat = m}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$get_X2()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_df_X2"></a>}}
\subsection{Method \code{get_df_X2()}}{
Returns degrees of freedom of Wald chi-square statistic.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$get_df_X2()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_p_X2"></a>}}
\subsection{Method \code{get_p_X2()}}{
Returns p-value for Wald chi-square statistic.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$get_p_X2()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Formatted printing of the \code{Contrast} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed on to \code{\link{print_ests}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
