% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FamData.R
\docType{class}
\name{FamData-class}
\alias{FamData-class}
\alias{FamData}
\title{Family data for mixed modeling}
\description{
A \code{R6} class that stores family data and a kinship matrix. Methods of
this class can then be used to fit genetic mixed models.
}
\section{Usage}{

\strong{Constructors}
\preformatted{fdata <- FamData$new(data, family_id, indiv_id, proband, sex,
  maternal_id, paternal_id, mzgrp, dzgrp)}
\preformatted{fdata <- FamData$new(data, family_id, indiv_id, proband, phi)}
\strong{Accessors}
\preformatted{fdata$get_data()}
\preformatted{fdata$get_data_name()}
\preformatted{fdata$get_phi()}
\preformatted{fdata$get_var_map()}
\strong{Model Fitting}
\preformatted{fdata$lmm(formula, ...)}
\strong{Plotting}
\preformatted{fdata$plot_pedigree(famid)}
}

\section{Arguments}{

\strong{Constructors}
\describe{
  \item{\code{data}}{A \code{data.frame} or \code{data.table}
    containing all phenotypes and genetic variables of interest.}
  \item{\code{family_id}}{The name of a numeric or character column in
    \code{data} containing a family identifier.}
  \item{\code{indiv_id}}{The name of a numeric or character column in
     \code{data} containing a individual identifier.}
  \item{\code{proband}}{The name of a numeric column in \code{data}
    containing 1 for probands and 0 otherwise. \code{NA} is not permitted.}
  \item{\code{sex}}{The name of a numeric column in \code{data} containing
    the individual's sex (1 = "male", 2 = "female") or \code{NA} if
    unknown.}
  \item{\code{maternal_id}}{The name of a column in \code{data}
    of the same type as \code{indiv_id} containing the \code{indiv_id} of
    the individual's mother or 0 (numeric), "" (character), or \code{NA} if
    a founder. See documentation for \code{\link[kinship2]{pedigree}}.}
  \item{\code{paternal_id}}{The name of a column in \code{data}
    of the same type as \code{indiv_id} containing the \code{indiv_id} of
    the individual's father or or 0 (numeric), "" (character), or \code{NA}
    if a founder. See documentation for \code{\link[kinship2]{pedigree}}.}
  \item{\code{mzgrp}}{The name of a numeric, character, or factor column in
    \code{data} containing the same value for all members of a monozygotic
    twin group. \code{NA} should be used for all individuals who are not
    monozygotic twins.}
  \item{\code{dzgrp}}{The name of a numeric, character, or factor column in
    \code{data} containing the same value for all members of a dizygotic
    twin group. \code{NA} should be used for all individuals who are not
    dizygotic twins.}
  \item{\code{phi}}{A matrix coercible to a
    \code{\link[Matrix:dsCMatrix-class]{dsCMatrix}} containing the
    pairwise kinship coefficients for individuals. Note that
    \code{phi[i, j]} must contain the kinship coefficient between the
    individuals in \code{data[i, ]} and \code{data[j, ]}.}
}
\strong{Model Fitting}
\describe{
  \item{\code{formula}}{A \code{formula} object describing the mean model
    for a phenotype.}
  \item{\code{...}}{Additional arguments for \code{\link[stats]{nlminb}}.}
}
\strong{Plotting}
\describe{
  \item{\code{famid}}{A value of \code{family_id} from the original data
    set.}
}
}

\section{Methods}{

\strong{Constructors}
\describe{
  \item{\code{$new()}}{Constructs a new instance of this class. See Usage
    for possible constructor signatures.}
}
\strong{Accessors}
\describe{
  \item{\code{$get_data()}}{Returns data member.}
  \item{\code{$get_data_name()}}{Returns name of input data set.}
  \item{\code{$get_phi()}}{Returns kinship matrix member.}
  \item{\code{$get_var_map()}}{Returns list mapping variable names in input
    data set to those in data member.}
}
\strong{Model Fitting}
\describe{
  \item{\code{$lmm()}}{Fits a linear mixed model. Returns a \code{list}
  object with the following contents:
  \describe{
    \item{\code{converge}}{Convergence code from
      \code{\link[stats]{nlminb}}. 0 is successful convergence.}
    \item{\code{message}}{Convergence message from
      \code{\link[stats]{nlminb}}.}
    \item{\code{max_grad}}{Maximum gradient component at solution.}
    \item{\code{deviance}}{-2 loglikelihood at solution.}
    \item{\code{parm_ests}}{\code{matrix} of parameter estimates, standard
      errors, and Wald tests/p-values.}
    \item{\code{vc_pl_cis}}{\code{matrix} of profile likelihood confidence
      intervals for variance components. \code{NA} if confidence level is
      not attainable or something went wrong in likelihood profiling.}
    \item{\code{lrts}}{\code{matrix} of likelihood ratio test results. Row
      names denote the null hypotheses, and columns indicate whether the
      null maximization converged, the X2 statistic, and the p-value
      calculated according to the approach of Self and Liang (1987).}
  }}
}
\strong{Plotting}
\describe{
  \item{\code{$plot_pedigree()}}{For objects initialized with pedigree data,
    allows plotting of individual pedigrees using \pkg{kinship2}.}
}
}

\references{
Self SG, Liang K-Y. 1987. Asymptotic Properties of Maximum Likelihood
  Estimators and Likelihood Ratio Tests Under Nonstandard Conditions.
  \emph{J Am Stat Assoc} 82 (398): 605-10.
  \url{https://doi.org/10.2307/2289471}.
}
