% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FamLMMFit.R
\name{FamLMMFit}
\alias{FamLMMFit}
\title{A \code{FamModelFit} subclass specialized for linear mixed models}
\description{
A specialized subclass of the
  \code{\link[FamModel]{FamModelFit}} abstract class for linear mixed model
  results. Methods of this subclass can be used to print model results in a
  nice format, access model objects, perform inference on general contrasts,
  and perform likelihood ratio tests on variance parameters.
}
\references{
Self SG, Liang K-Y. 1987. Asymptotic Properties of Maximum Likelihood
  Estimators and Likelihood Ratio Tests Under Nonstandard Conditions.
  \emph{J Am Stat Assoc} 82 (398): 605-10.
  \url{https://doi.org/10.2307/2289471}.
}
\section{Super class}{
\code{\link[FamModel:FamModelFit]{FamModel::FamModelFit}} -> \code{FamLMMFit}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FamLMMFit$new()}}
\item \href{#method-get_formula}{\code{FamLMMFit$get_formula()}}
\item \href{#method-get_objfun}{\code{FamLMMFit$get_objfun()}}
\item \href{#method-get_objfun_par_names}{\code{FamLMMFit$get_objfun_par_names()}}
\item \href{#method-get_r_cholesky}{\code{FamLMMFit$get_r_cholesky()}}
\item \href{#method-get_r_cond}{\code{FamLMMFit$get_r_cond()}}
\item \href{#method-get_r_pop}{\code{FamLMMFit$get_r_pop()}}
\item \href{#method-get_sdreport}{\code{FamLMMFit$get_sdreport()}}
\item \href{#method-print}{\code{FamLMMFit$print()}}
\item \href{#method-h2_a_lrts}{\code{FamLMMFit$h2_a_lrts()}}
\item \href{#method-clone}{\code{FamLMMFit$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="FamModel" data-topic="FamModelFit" data-id="contrast">}\href{../../FamModel/html/FamModelFit.html#method-contrast}{\code{FamModel::FamModelFit$contrast()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="FamModel" data-topic="FamModelFit" data-id="get_data">}\href{../../FamModel/html/FamModelFit.html#method-get_data}{\code{FamModel::FamModelFit$get_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="FamModel" data-topic="FamModelFit" data-id="get_optres">}\href{../../FamModel/html/FamModelFit.html#method-get_optres}{\code{FamModel::FamModelFit$get_optres()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="FamModel" data-topic="FamModelFit" data-id="get_theta_hat">}\href{../../FamModel/html/FamModelFit.html#method-get_theta_hat}{\code{FamModel::FamModelFit$get_theta_hat()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="FamModel" data-topic="FamModelFit" data-id="get_V_theta_hat">}\href{../../FamModel/html/FamModelFit.html#method-get_V_theta_hat}{\code{FamModel::FamModelFit$get_V_theta_hat()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Constructs a new instance of this class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamLMMFit$new(data, formula, objfun, optres)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The \code{\link{FamData}} object that produced this model
fit.}

\item{\code{formula}}{The \code{\link[Formula]{Formula}} object that describes
the model.}

\item{\code{objfun}}{A \code{\link[TMB]{MakeADFun}} object that has already been
run through an optimizer. This object is expected to have a "type"
attribute of "TMB".}

\item{\code{optres}}{A \code{list} from the optimizer function detailing
results. This list is expected to have an attribute "opttype" that
contains a string identifying the function used for optimization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_formula"></a>}}
\subsection{Method \code{get_formula()}}{
Returns the \code{\link[Formula]{Formula}}
  \code{\link[Formula]{Formula}} object used to construct the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamLMMFit$get_formula()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_objfun"></a>}}
\subsection{Method \code{get_objfun()}}{
Returns \code{\link[TMB]{MakeADFun}} object used to
  construct the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamLMMFit$get_objfun()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_objfun_par_names"></a>}}
\subsection{Method \code{get_objfun_par_names()}}{
Returns a vector of names derived from the original
  formula that map to each element of \code{objfun$par}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamLMMFit$get_objfun_par_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_r_cholesky"></a>}}
\subsection{Method \code{get_r_cholesky()}}{
Returns the \emph{conditional} residuals scaled by the
  the inverse Cholesky factor of the covariance matrix of the
  conditional distribution of the family members given the proband.
  These are defined for family members only. If the model is correct,
  these should approximate independent standard normal variables in
  large samples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamLMMFit$get_r_cholesky()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_r_cond"></a>}}
\subsection{Method \code{get_r_cond()}}{
Returns the \emph{conditional} residuals, defined as
  \code{y - X * b - d_pr}, where \code{X * b + d_pr} is the mean in
  the family member conditional upon the data in the proband. These
  are defined for family members only.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamLMMFit$get_r_cond()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_r_pop"></a>}}
\subsection{Method \code{get_r_pop()}}{
Returns the \emph{population} residuals, defined as
  \code{y - X * b}, for everyone, including the proband.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamLMMFit$get_r_pop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sdreport"></a>}}
\subsection{Method \code{get_sdreport()}}{
Returns the \code{\link[TMB]{sdreport}} object produced
  from the \code{\link[TMB]{MakeADFun}} object used to construct the
  instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamLMMFit$get_sdreport()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Prints contents of \code{FamLMMFit} objects with nice
  formatting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamLMMFit$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed on to \code{\link{print_ests}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-h2_a_lrts"></a>}}
\subsection{Method \code{h2_a_lrts()}}{
Perform likelihood ratio test(s) for no polygenic effect(s)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamLMMFit$h2_a_lrts(print = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{print}}{If \code{TRUE} (default), prints the likelihood ratio test
results in a nice format.}

\item{\code{...}}{Additional parameters to pass to the optimization function
(currently \code{\link{nlminb}}).}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Tests that each narrow-sense heritability parameter is zero
  individually using a likelihood ratio test. As these null hypotheses
  are on the boundary of the parameter space, a 50:50 mixture of
  chi-square(0) and chi-square(1) is used for the p-value (see Self and
  Liang).
}

\subsection{Returns}{
A \code{data.table} containing likelihood ratio test results,
  invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamLMMFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
