% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FamModelFit.R
\name{FamModelFit}
\alias{FamModelFit}
\title{Abstract class for results of a fitted \code{FamModel} model}
\description{
An abstract R6 class defining the pattern for all
  \code{\link{FamModelFit}} subclasses created by fitting a model to a
  \code{\link{FamData}} object using one of its model-fitting methods.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FamModelFit$new()}}
\item \href{#method-get_data}{\code{FamModelFit$get_data()}}
\item \href{#method-get_optres}{\code{FamModelFit$get_optres()}}
\item \href{#method-get_theta_hat}{\code{FamModelFit$get_theta_hat()}}
\item \href{#method-get_V_theta_hat}{\code{FamModelFit$get_V_theta_hat()}}
\item \href{#method-print}{\code{FamModelFit$print()}}
\item \href{#method-contrast}{\code{FamModelFit$contrast()}}
\item \href{#method-clone}{\code{FamModelFit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Stub constructor to prevent instantiation of this abstract
  class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamModelFit$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\subsection{Method \code{get_data()}}{
Returns the \code{\link{FamData}} object that produced this
  model fit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamModelFit$get_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_optres"></a>}}
\subsection{Method \code{get_optres()}}{
Returns \code{list} of optimization results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamModelFit$get_optres()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_theta_hat"></a>}}
\subsection{Method \code{get_theta_hat()}}{
Returns \code{theta_hat}, the vector of parameter
  estimates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamModelFit$get_theta_hat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_V_theta_hat"></a>}}
\subsection{Method \code{get_V_theta_hat()}}{
Returns \code{V_theta_hat}, the covariance matrix of the
  parameter estimates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamModelFit$get_V_theta_hat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Formatted printing of the \code{FamModelFit} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamModelFit$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed on to \code{\link{print_ests}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-contrast"></a>}}
\subsection{Method \code{contrast()}}{
Create a new \code{\link{Contrast}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamModelFit$contrast(L_mat, m)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{L_mat}}{A contrast vector (1 df) or \code{matrix} (>1 df) containing
one contrast in each row.}

\item{\code{m}}{An optional vector containing the null value for each contrast.
Will be set to the zero vector of length \code{nrow(L_mat)} if not
specified.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link{Contrast}} object for the specified arguments and
  this model fit.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamModelFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
