% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{print_ests}
\alias{print_ests}
\title{Print parameter estimates}
\usage{
print_ests(theta_hat, V_theta_hat, trans, ...)
}
\arguments{
\item{theta_hat}{A named vector of parameter estimates.}

\item{V_theta_hat}{The covariance matrix of the parameter estimates in
\code{theta_hat}.}

\item{trans}{A function accepting a single argument (e.g., \code{exp}) that
is used to transform parameter estimates and confidence intervals. All
tests and standard errors are still on the untransformed scale.}

\item{...}{
  Arguments passed on to \code{\link[=make_coef_mat]{make_coef_mat}}, \code{\link[stats:printCoefmat]{stats::printCoefmat}}
  \describe{
    \item{\code{alpha}}{The alpha level for the confidence intervals; default 0.05.}
    \item{\code{SEs}}{\code{TRUE} (default) if standard errors should be reported.}
    \item{\code{CIs}}{\code{TRUE} (default) if \eqn{100(1 - \alpha)}\% CIs should be
reported.}
    \item{\code{tests}}{\code{TRUE}(default) if Wald test statistics and p-values based
on the normal distribution should be reported.}
    \item{\code{digits}}{minimum number of significant digits to be used for
    most numbers.}
    \item{\code{signif.stars}}{logical; if \code{TRUE}, P-values are additionally
    encoded visually as \sQuote{significance stars} in order to help scanning
    of long coefficient tables.  It defaults to the
    \code{show.signif.stars} slot of \code{\link{options}}.}
    \item{\code{signif.legend}}{logical; if \code{TRUE}, a legend for the
    \sQuote{significance stars} is printed provided \code{signif.stars =
      TRUE}.}
    \item{\code{dig.tst}}{minimum number of significant digits for the test statistics,
    see \code{tst.ind}.}
    \item{\code{zap.ind}}{indices (integer) of column numbers which should be
    formatted by \code{\link{zapsmall}}, i.e., by \sQuote{zapping} values
    close to 0.}
    \item{\code{P.values}}{logical or \code{NULL}; if \code{TRUE}, the last
    column of \code{x} is formatted by \code{\link{format.pval}} as P
    values.  If \code{P.values = NULL}, the default, it is set to
    \code{TRUE} only if \code{\link{options}("show.coef.Pvalue")} is
    \code{TRUE} \emph{and} \code{x} has at least 4 columns \emph{and}
    the last column name of \code{x} starts with \code{"Pr("}.}
    \item{\code{has.Pvalue}}{logical; if \code{TRUE}, the last column of \code{x}
    contains P values; in that case, it is printed if and only if
    \code{P.values} (above) is true.}
    \item{\code{eps.Pvalue}}{number, ..}
    \item{\code{na.print}}{a character string to code \code{\link{NA}} values in
    printed output.}
    \item{\code{quote}}{further arguments passed to
    \code{\link{print.default}}.}
    \item{\code{right}}{further arguments passed to
    \code{\link{print.default}}.}
  }}
}
\description{
Prints estimates, standard errors, \eqn{100(1 - \alpha)}\% CIs, test
statistics, and p-values to the console given a vector of estimates and their
covariance matrix. May also specify a function to transform the estimates and
95\% CIs.  Uses \code{\link[=make_coef_mat]{make_coef_mat()}} for underlying calculations.
}
