% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contrast.R
\name{Contrast}
\alias{Contrast}
\title{Contrast estimated from an object inheriting from \code{FamModelFit}}
\description{
An R6 class that stores results for a contrast of the form
\eqn{L \theta - m} estimated using a specified object inheriting from
\code{\link{FamModelFit}}. Wald tests and confidence intervals for individual
rows of the contrast matrix as well as an overall Wald chi-square test of
the null hypothesis \eqn{L \theta - m = 0} are provided.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Contrast$new()}}
\item \href{#method-get_model_fit}{\code{Contrast$get_model_fit()}}
\item \href{#method-get_L_mat}{\code{Contrast$get_L_mat()}}
\item \href{#method-get_m}{\code{Contrast$get_m()}}
\item \href{#method-get_L_theta_hat_m}{\code{Contrast$get_L_theta_hat_m()}}
\item \href{#method-get_V_L_theta_hat}{\code{Contrast$get_V_L_theta_hat()}}
\item \href{#method-get_X2}{\code{Contrast$get_X2()}}
\item \href{#method-get_df_X2}{\code{Contrast$get_df_X2()}}
\item \href{#method-get_p_X2}{\code{Contrast$get_p_X2()}}
\item \href{#method-print}{\code{Contrast$print()}}
\item \href{#method-clone}{\code{Contrast$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructs a new instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$new(model_fit, L_mat, m)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_fit}}{An object inheriting from \code{\link{FamModelFit}}.}

\item{\code{L_mat}}{A contrast vector (1 df) or \code{matrix} (>1 df) containing
one contrast in each row. The contrast vector must have a number of
elements equal to the number of model parameters. The contrast matrix
must be of full row rank and have a number of columns equal the number
of model parameters.}

\item{\code{m}}{An optional vector containing the null value for each contrast.
Will be set to the zero vector of length \code{nrow(L_mat)} if not
specified.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_model_fit"></a>}}
\if{latex}{\out{\hypertarget{method-get_model_fit}{}}}
\subsection{Method \code{get_model_fit()}}{
Returns \code{model_fit} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$get_model_fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_L_mat"></a>}}
\if{latex}{\out{\hypertarget{method-get_L_mat}{}}}
\subsection{Method \code{get_L_mat()}}{
Returns \code{L_mat}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$get_L_mat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_m"></a>}}
\if{latex}{\out{\hypertarget{method-get_m}{}}}
\subsection{Method \code{get_m()}}{
Returns \code{m}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$get_m()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_L_theta_hat_m"></a>}}
\if{latex}{\out{\hypertarget{method-get_L_theta_hat_m}{}}}
\subsection{Method \code{get_L_theta_hat_m()}}{
Returns \eqn{L \hat{\theta} - m}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$get_L_theta_hat_m()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_V_L_theta_hat"></a>}}
\if{latex}{\out{\hypertarget{method-get_V_L_theta_hat}{}}}
\subsection{Method \code{get_V_L_theta_hat()}}{
Returns \eqn{L \hat{V}(\hat{\theta}) L^{'}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$get_V_L_theta_hat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_X2"></a>}}
\if{latex}{\out{\hypertarget{method-get_X2}{}}}
\subsection{Method \code{get_X2()}}{
Returns Wald chi-square statistic for null hypothesis
\eqn{L \theta - m = 0}, which is \eqn{(L \hat{\theta} - m)^{'}
  [L \hat{V}(\hat{\theta}) L^{'}]^{-1} (L \hat{\theta} - m)}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$get_X2()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_df_X2"></a>}}
\if{latex}{\out{\hypertarget{method-get_df_X2}{}}}
\subsection{Method \code{get_df_X2()}}{
Returns degrees of freedom of Wald chi-square statistic,
which is the rank of \eqn{L}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$get_df_X2()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_p_X2"></a>}}
\if{latex}{\out{\hypertarget{method-get_p_X2}{}}}
\subsection{Method \code{get_p_X2()}}{
Returns p-value for Wald chi-square statistic.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$get_p_X2()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Formatted printing of the \code{Contrast} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed on to \code{\link[=print_ests]{print_ests()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contrast$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
