% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FamData.R
\name{FamData}
\alias{FamData}
\title{Family data for modeling}
\description{
An R6 class that stores family data and a kinship matrix.
}
\details{
Methods of this class can be used to fit regression models such as
polygenic mixed models. The class currently supports only family data with
one observation per subject and will check that this requirement is met.

Note that the private \code{data} member of the instance (returned by the
\href{#method-get_data}{\code{FamData$get_data()}} method) will always be sorted by
ascending \code{family_id} and ascending \code{indiv_id} within \code{family_id}
regardless of the sort order of the input data set. Also, pedigree columns
in the input data set will be renamed in the \code{data} member; this mapping
can be returned using the \href{#method-get_var_map}{\code{FamData$get_var_map()}}
method.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FamData$new()}}
\item \href{#method-get_data}{\code{FamData$get_data()}}
\item \href{#method-get_data_name}{\code{FamData$get_data_name()}}
\item \href{#method-get_consang}{\code{FamData$get_consang()}}
\item \href{#method-get_phi}{\code{FamData$get_phi()}}
\item \href{#method-get_var_map}{\code{FamData$get_var_map()}}
\item \href{#method-print}{\code{FamData$print()}}
\item \href{#method-lmm}{\code{FamData$lmm()}}
\item \href{#method-plot_pedigree}{\code{FamData$plot_pedigree()}}
\item \href{#method-clone}{\code{FamData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructs a new instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamData$new(data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A \code{data.frame} or \code{\link{data.table}} containing all
phenotypes and genetic variables of interest.}

\item{\code{...}}{The arguments required for a particular method signature;
see below for detailed descriptions.}

\item{\code{family_id}}{Character string containing the name of a numeric or
character column in \code{data} containing a family identifier.}

\item{\code{indiv_id}}{Character string containing the name of a numeric or
character column in \code{data} containing a individual identifier.}

\item{\code{proband}}{Character string containing the name of a numeric column
in \code{data} containing 1 for probands and 0 otherwise. \code{NA} is  not
permitted.}

\item{\code{sex}}{Character string containing the name of a numeric column in
\code{data} containing the individual's sex (1 = "male", 2 = "female")
or \code{NA} if unknown.}

\item{\code{maternal_id}}{Character string containing the name of a column in
\code{data} of the same type as \code{indiv_id} containing the \code{indiv_id} of the
individual's mother or 0 (numeric), "" (character), or \code{NA} if a
founder. See documentation for \code{\link[kinship2:pedigree]{kinship2::pedigree}}.}

\item{\code{paternal_id}}{Character string containing the name of a column in
\code{data} of the same type as \code{indiv_id} containing the \code{indiv_id} of the
individual's father or or 0 (numeric), "" (character), or \code{NA} if a
founder. See documentation for \code{\link[kinship2:pedigree]{kinship2::pedigree}}.}

\item{\code{mzgrp}}{Character string containing the name of a numeric,
character, or factor column in \code{data} containing the same value for all
members of a monozygotic twin group. \code{NA} should be used for all
individuals who are not monozygotic twins.}

\item{\code{dzgrp}}{Character string containing the name of a numeric,
character, or factor column in \code{data} containing the same value for all
members of a dizygotic twin group. \code{NA} should be used for all
individuals who are not dizygotic twins.}

\item{\code{phi}}{A matrix coercible to a \code{\link[Matrix:dsCMatrix-class]{Matrix::dsCMatrix}} containing
the pairwise kinship coefficients for individuals. Note that
\code{phi[i, j]} must contain the kinship coefficient between the individuals
in \code{data[i, ]} and \code{data[j, ]}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
All constructor arguments other than \code{data}
\emph{must be named}. There are two possible constructor signatures:\preformatted{FamData$new(data, family_id, indiv_id, proband, sex,
            maternal_id, paternal_id, mzgrp, dzgrp)

FamData$new(data, family_id, indiv_id, proband, phi)
}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
Returns a copy of data member (to prevent accidental
modification of data member \code{\link{data.table}} by reference).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamData$get_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_name"></a>}}
\if{latex}{\out{\hypertarget{method-get_data_name}{}}}
\subsection{Method \code{get_data_name()}}{
Returns name of input \code{data.frame} or \code{\link{data.table}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamData$get_data_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_consang"></a>}}
\if{latex}{\out{\hypertarget{method-get_consang}{}}}
\subsection{Method \code{get_consang()}}{
Returns character vector containing family IDs in which
consanguinity was found (based on individual self-kinship coefficient
greater than 0.5).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamData$get_consang()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_phi"></a>}}
\if{latex}{\out{\hypertarget{method-get_phi}{}}}
\subsection{Method \code{get_phi()}}{
Returns kinship matrix member.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamData$get_phi()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_var_map"></a>}}
\if{latex}{\out{\hypertarget{method-get_var_map}{}}}
\subsection{Method \code{get_var_map()}}{
Returns list mapping variable names in input data set in
\code{data} argument to those in data member.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamData$get_var_map()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints information about contents of \code{FamData} objects
with nice formatting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamData$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lmm"></a>}}
\if{latex}{\out{\hypertarget{method-lmm}{}}}
\subsection{Method \code{lmm()}}{
Fits a linear mixed model to ascertained families
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamData$lmm(formula, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{A \code{\link[Formula:Formula]{Formula::Formula()}} object describing the model for a
phenotype. The formula is of the form \code{y ~ mean | group}, where \code{y}
(required) is the outcome, \code{mean} (optional) specifies the mean model
that includes an intercept by default, and \verb{| group} (optional)
specifies a factor term that can be used assign families to homogeneous
groups with different variance components. Multiple variables can be
combined into a single factor using \code{:}.}

\item{\code{...}}{Additional parameters to pass to the \code{control} list for
\code{\link[=optim]{optim()}} with \code{method = "L-BGFS-B"}. Note that \code{parscale} and
\code{fnscale} cannot be modified.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See \code{vignette("linear_mixed_models")} for background,
implementation details, and references.
}

\subsection{Returns}{
Returns a \code{\link{FamLMMFit}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_pedigree"></a>}}
\if{latex}{\out{\hypertarget{method-plot_pedigree}{}}}
\subsection{Method \code{plot_pedigree()}}{
For objects initialized with pedigree data, allows plotting
of individual pedigrees using \code{\link[kinship2:plot.pedigree]{kinship2::plot.pedigree}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamData$plot_pedigree(famid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{famid}}{Argument matching the value of a single family in the
\code{family_id} column used in the constructor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
