% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FamModelFit.R
\name{FamModelFit}
\alias{FamModelFit}
\title{Abstract class for results of a fitted \code{FamModel} model}
\description{
An abstract R6 class defining the pattern for all \code{FamModelFit}
subclasses created by fitting a model to a \code{\link{FamData}} object using one
of its model-fitting methods. This class provides only a default
constructor without any initializer.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_data}{\code{FamModelFit$get_data()}}
\item \href{#method-get_optres}{\code{FamModelFit$get_optres()}}
\item \href{#method-get_theta_hat}{\code{FamModelFit$get_theta_hat()}}
\item \href{#method-get_V_theta_hat}{\code{FamModelFit$get_V_theta_hat()}}
\item \href{#method-print}{\code{FamModelFit$print()}}
\item \href{#method-contrast}{\code{FamModelFit$contrast()}}
\item \href{#method-clone}{\code{FamModelFit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
Returns the \code{\link{FamData}} object that produced this model
fit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamModelFit$get_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_optres"></a>}}
\if{latex}{\out{\hypertarget{method-get_optres}{}}}
\subsection{Method \code{get_optres()}}{
Returns \code{list} of optimization results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamModelFit$get_optres()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_theta_hat"></a>}}
\if{latex}{\out{\hypertarget{method-get_theta_hat}{}}}
\subsection{Method \code{get_theta_hat()}}{
Returns \eqn{\hat{\theta}}, the numeric vector of parameter
estimates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamModelFit$get_theta_hat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_V_theta_hat"></a>}}
\if{latex}{\out{\hypertarget{method-get_V_theta_hat}{}}}
\subsection{Method \code{get_V_theta_hat()}}{
Returns the numeric \code{matrix} \eqn{\hat{V}(\hat{\theta})},
the estimated covariance matrix of the parameter estimates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamModelFit$get_V_theta_hat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Formatted printing of the \code{FamModelFit} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamModelFit$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed on to \code{\link[=print_ests]{print_ests()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-contrast"></a>}}
\if{latex}{\out{\hypertarget{method-contrast}{}}}
\subsection{Method \code{contrast()}}{
Create a new \code{Contrast} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamModelFit$contrast(L_mat, m)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{L_mat}}{A contrast vector (1 df) or \code{matrix} (>1 df) containing
one contrast in each row. The contrast vector must have a number of
elements equal to the number of model parameters. The contrast matrix
must be of full row rank and have a number of columns equal the number
of model parameters.}

\item{\code{m}}{An optional vector containing the null value for each contrast.
Will be set to the zero vector of length \code{nrow(L_mat)} if not
specified.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link{Contrast}} object for the specified arguments and this
model fit.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamModelFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
